Refactor Sample import and clean up code
- Moved Sample import from Profiles to simulation module
- Cleaned up unused imports in job files
- Updated references in unit tests accordingly
