--- src/easydiffraction/Profiles/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-FileCopyrightText: 2024 EasyDiffraction contributors
-# SPDX-License-Identifier: BSD-3-Clause
-# © 2021-2024 Contributors to the EasyDiffraction project <https://github.com/EasyScience/EasyDiffraction>
