+++ src/easydiffraction/job/experiment/data_container.py
@@ -5,7 +5,7 @@
 from easyscience.Objects.core import ComponentSerializer
 
 from easydiffraction.job.experiment.experiment import Experiment
-from easydiffraction.Profiles.Sample import Sample
+from easydiffraction.job.experiment.simulation import Sample
 
 class DataContainer(ComponentSerializer):
     def __init__(self, sim_store: Sample, exp_store: Experiment):
